fix: 빌드 오류 해결

+++ src/actions/deleteReview.ts
@@ -2,7 +2,11 @@
 
 import { getFirestore } from "firebase-admin/firestore";
 
-export const deleteReview = async (id: string) => {
+export const deleteReview = async (id?: string) => {
+  if (!id) {
+    throw new Error("삭제할 리뷰 ID가 없습니다.");
+  }
+
   const db = getFirestore();
   const ref = db.collection("reviews").doc(id);
   await ref.delete();
